Issue #2 - Recompile with STACK_SIZE = 5000 in types.h (!!)
Seems like an absurdly low size but in limited testing this prevents call stack exceeded errors when debugging on Chrome. May need to experiment with this more and/or check various browser settings. For the purposes of this terminal demo performance is not a primary concern, but I have to think a performance-intensive application would want to increase this number to avoid the overhead of constantly running minor GC.